requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-js-asset==3.1.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
phonenumberslite==9.0.1
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
